Time: 2 ms (98.91%), Space: 83.9 MB (75.9%) - LeetHub

+++ 2110-number-of-smooth-descent-periods-of-a-stock/2110-number-of-smooth-descent-periods-of-a-stock.java
@@ -1,8 +1,8 @@
 class Solution {
     public long getDescentPeriods(int[] prices) {
         int n=prices.length;
-        int ans=1;
-        int len=1;
+        long ans=1;
+        long len=1;
             for(int i=1;i<n;i++){
             if(prices[i-1]-prices[i]==1){
                 len++;
